backend/requirements.txt
# Web Framework
fastapi>=0.109.0

# Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0          # async driver (FastAPI / asyncio)
psycopg2-binary>=2.9.9   # sync driver (Celery tasks, StateDBConnector)
# aiosqlite no longer required — kept as comment for SQLite fallback dev
# aiosqlite>=0.19.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Task Queue
celery[redis]>=5.3.6
redis>=5.0.0

# Deep Learning
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.37.0
ultralytics>=8.1.0

# Utilities
loguru>=0.7.2
python-dotenv>=1.0.0
Pillow>=10.2.0
opencv-python-headless>=4.9.0
pyyaml>=6.0.1
